backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
openai>=1.50.0
supabase>=2.10.0
tiktoken>=0.8.0
pypdf>=4.0.0
python-multipart>=0.0.9
python-docx>=1.1.0
structlog>=24.4.0
opentelemetry-sdk>=1.27.0
opentelemetry-exporter-otlp-proto-grpc>=1.27.0
opentelemetry-exporter-prometheus>=0.48b0
opentelemetry-instrumentation-fastapi>=0.48b0
opentelemetry-instrumentation-httpx>=0.48b0
opentelemetry-instrumentation-logging>=0.48b0
opentelemetry-exporter-otlp-proto-http>=1.27.0
